refactor(fake): consolidate test files into fake_test.go
- Merged 9 separate test files into single fake_test.go
- Combined unit tests and benchmark tests in one file
- Removed: basic_test.go, benchmark_test.go, example_test.go, extreme_bench_test.go, extreme_example_test.go, nano_bench_test.go, optimization_bench_test.go, performance_first_bench_test.go, verification_test.go
- Maintains all existing test functionality
- Follows Go naming convention: fake.go -> fake_test.go
